renderers as parameters
I would like to add a plaintext renderer that works out of the box with all the plugins.

My approach would be to add renderers as a parameter, so other projects can provide a different set of default rules. Alternatively one could set the default to `None` and then assign `DEFAULT_RENDERERS` later. While this is cleaner, the type signature would have to account for this `Union[Mapping[str, Render], NoneType]`.

Alternatively I would duplicate the whole function or would write write a renderer class with methods for all the token types of all the plugins.